requirements.txt
django-appconf==1.0.5
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
django-imagekit==4.1.0
django-model-utils==4.3.1
django-redis==5.3.0
django==4.2.3
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
mysqlclient==2.2.0
pilkit==2.0
pillow==10.0.0
pyjwt==2.8.0
pymysql==1.1.0
redis==4.6.0
gunicorn
supervisor
django-adminlte-ui
requests